+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
+### [Added]
+- Adds option to append additional details to changelog entries ~ picorion (2024-04-11)
 ### [Changed]
 - Moves the specification of the target file to the hook
 ### [Fixed]